Refactored getting similar movies to the MovieDetail package
