+++ app/src/main/java/me/androidbox/busbymovies/di/ApplicationComponent.java
@@ -6,6 +6,7 @@ import dagger.Component;
 import me.androidbox.busbymovies.data.MovieFavouritePresenterImp;
 import me.androidbox.busbymovies.moviedetails.MovieDetailModelImp;
 import me.androidbox.busbymovies.moviedetails.MovieDetailPresenterImp;
+import me.androidbox.busbymovies.moviedetails.MovieDetailViewContract;
 import me.androidbox.busbymovies.moviedetails.MovieDetailViewImp;
 import me.androidbox.busbymovies.movielist.MovieListModelImp;
 import me.androidbox.busbymovies.movielist.MovieListPresenterImp;
