+++ app/src/main/java/me/androidbox/busbymovies/movielist/MovieListPresenterContract.java
@@ -10,5 +10,4 @@ public interface MovieListPresenterContract<View extends MovieListViewContract> 
     void getPopularMovies();
     void getTopRatedMovies();
     void searchMovies(final String movieName, final int movieYear);
-    void getSimilarMovies(final int movieId);
 }
